+++ src/Asv.Mavlink/Protocol/V2/MavlinkV2MessageFactory.cs
@@ -24,14 +24,14 @@ using Asv.Mavlink.UnitTestMessage;
 
 namespace Asv.Mavlink;
 
-public class MavlinkV2MessageFactory : IProtocolMessageFactory<MavlinkMessage, ushort>
+public class MavlinkV2MessageFactory : IProtocolMessageFactory<MavlinkMessage, int>
 {
-    private readonly ImmutableDictionary<ushort,Func<MavlinkMessage>> _decoder;
+    private readonly ImmutableDictionary<int,Func<MavlinkMessage>> _decoder;
     public static MavlinkV2MessageFactory Instance { get; } = new();
 
     private MavlinkV2MessageFactory()
     {
-        var builder = ImmutableDictionary.CreateBuilder<ushort, Func<MavlinkMessage>>();
+        var builder = ImmutableDictionary.CreateBuilder<int, Func<MavlinkMessage>>();
         builder.RegisterMinimalDialect();
         builder.RegisterCommonDialect();
         builder.RegisterArdupilotmegaDialect();
@@ -54,12 +54,12 @@ public class MavlinkV2MessageFactory : IProtocolMessageFactory<MavlinkMessage, u
     }
 
     
-    public MavlinkMessage? Create(ushort id) => _decoder.TryGetValue(id, out var factory) ? factory() : null;
+    public MavlinkMessage? Create(int id) => _decoder.TryGetValue(id, out var factory) ? factory() : null;
 
     public ProtocolInfo Info => MavlinkV2Protocol.Info;
     
-    public IEnumerable<ushort> GetSupportedIds()
+    public IEnumerable<int> GetSupportedIds()
     {
-        return _decoder.Keys.Select(x=>(ushort)x);
+        return _decoder.Keys.Select(x=>(int)x);
     }
 }
